LPS-78208 Move FragmentItemSelectorWebKeys to internal.constants
