+++ modules/apps/web-experience/fragment/fragment-item-selector-web/src/main/resources/META-INF/resources/init.jsp
@@ -23,7 +23,7 @@ taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.fragment.item.selector.web.internal.FragmentItemSelectorWebKeys" %><%@
+<%@ page import="com.liferay.fragment.item.selector.web.internal.constants.FragmentItemSelectorWebKeys" %><%@
 page import="com.liferay.fragment.item.selector.web.internal.context.FragmentItemSelectorViewDisplayContext" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
